requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
aiohttp==3.7.4
    # via cloudfoundry-client
asgiref==3.4.1
    # via django
async-timeout==3.0.1
    # via aiohttp
attrs==21.2.0
    # via aiohttp
certifi==2021.5.30
    # via
    #   requests
    #   sentry-sdk
chardet==3.0.4
    # via aiohttp
charset-normalizer==2.0.4
    # via requests
cloudfoundry-client==1.20.4
    # via -r requirements.in
dj-database-url==0.5.0
    # via -r requirements.in
django==3.2.5
    # via
    #   -r requirements.in
    #   django-staff-sso-client
django-environ==0.4.5
    # via -r requirements.in
django-staff-sso-client==3.1.0
    # via -r requirements.in
geckoboard.py==1.0.0
    # via -r requirements.in
gunicorn==20.1.0
    # via -r requirements.in
idna==3.2
    # via
    #   requests
    #   yarl
multidict==5.1.0
    # via
    #   aiohttp
    #   yarl
oauth2-client==1.2.1
    # via cloudfoundry-client
oauthlib==3.1.1
    # via requests-oauthlib
polling2==0.4.6
    # via cloudfoundry-client
protobuf==3.17.3
    # via cloudfoundry-client
psycopg2-binary==2.9.1
    # via -r requirements.in
python-dateutil==2.8.2
    # via -r requirements.in
python-dotenv==0.19.0
    # via -r requirements.in
pytz==2021.1
    # via django
pyyaml==5.4.1
    # via cloudfoundry-client
requests==2.26.0
    # via
    #   -r requirements.in
    #   cloudfoundry-client
    #   geckoboard.py
    #   oauth2-client
    #   requests-oauthlib
requests-oauthlib==1.3.0
    # via django-staff-sso-client
sentry-sdk==1.3.1
    # via -r requirements.in
six==1.16.0
    # via
    #   protobuf
    #   python-dateutil
    #   websocket-client
sqlparse==0.4.1
    # via django
typing-extensions==3.10.0.0
    # via aiohttp
urllib3==1.26.6
    # via
    #   requests
    #   sentry-sdk
websocket-client==0.54.0
    # via cloudfoundry-client
whitenoise==5.3.0
    # via -r requirements.in
yarl==1.6.3
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# setuptools
